Save last accepted version and show license dialog only if version was not accepted before
Note that the current implementation works when running `npm run electron`. However, the
builded version for windows (`npm run electron-package-app-production-windows`) does not work.
The reason is that we currently ignore the node_modules folder for these build but it is now required
because of an added dependency. For a detailed discussion on that, see https://gitlab.fit.fraunhofer.de/cortado/cortado/-/merge_requests/5.