Move angular dependencies to devDependencies
See https://stackoverflow.com/questions/49072226/electron-angular-cli-how-to-package-dist-with-electron-packager for detailled information.
In general, the problem was that we added some electron dependencies but do not want to keep a whitelist for the dependencies.
Therefore, we now only use the electron dependencies as dependencies in the package.json file and do not ignore node_modules for the production build.
Note that angular uses webpack to bundle the application and we therefore do not want to add it's packages to the production node_modules folder.