ft-227: Hide tooltips on scroll
Just as a short summary: As described in the issue, the problem occurs for the variants.
One idea was to use an angular wrapper for bootstrap like ngbootstrap. However, this
does not solve the problem, as the problem occurs in the dynamically added svg elements
(which cannot be covered by angular directives etc.). So, one further workaround is to hide all
tooltips when the user scrolls. Regarding the performance, this seems to be acceptable.